requirements.txt
discord.py
motor
matplotlib
asyncio
pymongo